FAT: SimpleFileSlack: start write behind RAM slack
As [1] describes, the windows implementation for NTFS padds RAM slack
with zeros. It seems the linux implementation of FAT does the same
thing, so this fix will leave RAM slack untouched and starts now writing
into slack at the start of the next free sector.

Test case was a FAT12 image. I filled the dataregion with 0xff, ran
fsck over it to fix destroyed references to directories, mounted it and
wrote a small file (<512 Byte). In hex presentation it was visible that
the remaining free space in the last occupied sector was padded with
0x00.

[1]: https://www.scribd.com/document/6642954/Ntfs-Hidden-Data-Analysis