CLI,FileSlack:TEST: create test filesystems under /tmp
now the unittests create their test images under /tmp. This way we
speed up the setup and reduce disk io.

Therfore a change to create_testfs.sh was necessary to provide an
alternate directory, where the test images will be stored. This makes
the ramdisk option obsolete, because one can provide simply the path
to a mounted tmpfs.