Filesystem_detector: added unit tests
This required changes to create_testfs, as we needed an unsupported
filesystem to check for an UnknownFilesystemError. With this change I
introduced a third command line option to specify the filesystem type
the images are generated for. This way unittests can explicitly choose
which filesystems shall be generated and setup-time will be reduced, if
not all filesystem types are needed.