"create_testfs: copy test images from directory
Introduced an option to copy existing test images from a directory instead of
creating them again.

This works by creating a config file called `.create_testfs.conf` which
will be sourced, if the USECONFIG option is set while running
`create_testfs.sh`. If there is a variable called `copyfrom` with a path
defined, this directory will be used as a source location of existing
test images.

This way, we don't need to prompt for a password when running unit tests.