FAT: added DirEntry class to wrap DIR_ENTRY Struct
The DirEntry class provides methods for commonly used actions like
if it is a file, directory or dot entry and getting the filename.

With this change, we are also able to work with directory entries that
dont have a long filename. Instead the name field is used. If it is a
deleted file, the name field is reconstructed.

Also this commit introduces a LFNEntries class which is used to get the
concatinated lfn. With this class the sequence number of lfn entries is
also respected, which wasn't the case before this commit.